requirements.txt
fastapi==0.103.1
openai==0.28.0
uvicorn==0.23.2
jinja2==2.7.2
markupsafe==2.0.1
python-decouple==3.8
